requirements_advanced.txt
# Core Framework
streamlit==1.31.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AI/ML Core
torch==2.1.2
torchvision==0.16.2
transformers==4.37.0
accelerate==0.26.1
diffusers==0.25.1
optimum==1.16.2

# VLLM & Performance
vllm==0.3.0
ray==2.9.0

# Model Optimization
onnx==1.15.0
onnxruntime-gpu==1.17.0
tensorrt==8.6.1

# Image Processing & CV
opencv-python==4.9.0.80
pillow==10.2.0
numpy==1.24.3
albumentations==1.3.1
rembg==2.0.56

# ControlNet & Advanced
controlnet-aux==0.0.7
insightface==0.7.3
mediapipe==0.10.9

# Model Serving
triton-client==2.41.0
kserve==0.11.2

# MLOps & Monitoring
mlflow==2.10.0
tensorboard==2.15.1
prometheus-client==0.19.0

# API & HTTP
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0

# Image & Vision Models
timm==0.9.12
segment-anything==1.0
groundingdino-py==0.4.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
loguru==0.7.2
tqdm==4.66.1
huggingface-hub==0.20.2

# Database (optional)
redis==5.0.1
sqlalchemy==2.0.25

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
